push

+++ srcs/pf_width.c
@@ -6,7 +6,7 @@
 /*   By: thgermai <thgermai@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/03 13:37:50 by thgermai          #+#    #+#             */
-/*   Updated: 2019/12/04 11:15:56 by thgermai         ###   ########.fr       */
+/*   Updated: 2019/12/04 14:34:18 by thgermai         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ char		*fill_width_right(char *str, t_param *param)
 {
 	char	*temp;
 
-	if (ft_strlen(str) > param->width)
+	if ((int)ft_strlen(str) > param->width)
 		return (temp = ft_strdup(str));
 	if (!(temp = malloc(sizeof(char) * param->width)))
 		return (NULL);
@@ -30,7 +30,7 @@ char		*fill_width_left(char *str, t_param *param)
 {
 	char	*temp;
 
-	if (ft_strlen(str) > param->width)
+	if ((int)ft_strlen(str) > param->width)
 		return (temp = ft_strdup(str));
 	if (!(temp = malloc(sizeof(char) * param->width)))
 		return (NULL);
